CastorEDC.Common.Studies.Users (ExCastorEDC v0.3.0-dev)

Provides access to user related endpoints within the study context

Functions

find(client, study_id, user_id)

Find a specific user that has access to the study

More info

list(client, study_id)

List all users that have access to the given study

More info