CastorEDC.DataCollection.DataEntry.ReportDataPoint (ExCastorEDC v0.2.0-dev)

Provides access to report data point related endpoints

Functions

create(client, study_id, record_id, report_instance_id, field_id, body)

Create a new data point in a specific report instance for a record

More info

find(client, study_id, record_id, report_instance_id, field_id, params \\ [])

Find a data point for a record in a specific report instance

More info

list(client, study_id, record_id, report_instance_id, params \\ [])

List all collected report instance data points for a record

More info