CastorEDC.DataCollection.DataEntry.StudyDataPoint (ExCastorEDC v0.4.0)

Provides access to study data point related endpoints

Functions

create(client, study_id, participant_id, field_id, body)

Create a new data point for a participant

More info

find(client, study_id, participant_id, field_id, params \\ [])

Find collected study data point for a participant in a field

More info

list(client, study_id, participant_id, params \\ [])

List all collected study data points for a participant

More info